v 1.8
Tipo primitivo da variável valor alterado para float

+++ Source Code Inspection/src/br/calebe/ticketmachine/core/PapelMoeda.java
@@ -6,10 +6,10 @@ package br.calebe.ticketmachine.core;
  */
 public class PapelMoeda {
 
-    protected int valor;
+    protected float valor;
     protected int quantidade;
 
-    public PapelMoeda(int valor, int quantidade) {
+    public PapelMoeda(float valor, int quantidade) {
         this.valor = valor;
         this.quantidade = quantidade;
     }
